Move NewsCard unit tests
