+++ WordPress/src/test/java/org/wordpress/android/viewmodel/reader/ReaderPostListViewModelTest.kt
@@ -4,8 +4,6 @@ import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Observer
 import com.nhaarman.mockitokotlin2.any
-import com.nhaarman.mockitokotlin2.argThat
-import com.nhaarman.mockitokotlin2.inOrder
 import com.nhaarman.mockitokotlin2.times
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
@@ -28,15 +26,12 @@ import org.wordpress.android.models.ReaderTag
 import org.wordpress.android.models.ReaderTagType.BOOKMARKED
 import org.wordpress.android.models.news.NewsItem
 import org.wordpress.android.ui.news.NewsManager
-import org.wordpress.android.ui.news.NewsTracker
-import org.wordpress.android.ui.news.NewsTracker.NewsCardOrigin.READER
-import org.wordpress.android.ui.news.NewsTrackerHelper
 import org.wordpress.android.ui.prefs.AppPrefsWrapper
 import org.wordpress.android.ui.reader.ReaderSubsActivity
 import org.wordpress.android.ui.reader.reblog.NoSite
 import org.wordpress.android.ui.reader.reblog.PostEditor
-import org.wordpress.android.ui.reader.reblog.Unknown
 import org.wordpress.android.ui.reader.reblog.SitePicker
+import org.wordpress.android.ui.reader.reblog.Unknown
 import org.wordpress.android.ui.reader.services.update.ReaderUpdateLogic.UpdateTask
 import org.wordpress.android.ui.reader.subfilter.ActionType.OpenLoginPage
 import org.wordpress.android.ui.reader.subfilter.ActionType.OpenSubsAtPage
@@ -59,17 +54,13 @@ class ReaderPostListViewModelTest {
     @JvmField
     val rule = InstantTaskExecutorRule()
 
-    @Mock private lateinit var newsManager: NewsManager
     @Mock private lateinit var observer: Observer<NewsItem>
     @Mock private lateinit var item: NewsItem
     /**
      * First tag for which the card was shown.
      */
     @Mock private lateinit var initialTag: ReaderTag
-    @Mock private lateinit var otherTag: ReaderTag
     @Mock private lateinit var savedTag: ReaderTag
-    @Mock private lateinit var newsTracker: NewsTracker
-    @Mock private lateinit var newsTrackerHelper: NewsTrackerHelper
     @Mock private lateinit var appPrefsWrapper: AppPrefsWrapper
     @Mock private lateinit var subfilterListItemMapper: SubfilterListItemMapper
     @Mock private lateinit var eventBusWrapper: EventBusWrapper
@@ -82,7 +73,6 @@ class ReaderPostListViewModelTest {
 
     @Before
     fun setUp() {
-        whenever(newsManager.newsItemSource()).thenReturn(liveData)
         whenever(savedTag.tagTitle).thenReturn("tag-title")
         val tag = Tag(
                 tag = savedTag,
@@ -94,9 +84,6 @@ class ReaderPostListViewModelTest {
         whenever(subfilterListItemMapper.fromJson(any(), any(), any())).thenReturn(tag)
 
         viewModel = ReaderPostListViewModel(
-                newsManager,
-                newsTracker,
-                newsTrackerHelper,
                 TEST_DISPATCHER,
                 TEST_DISPATCHER,
                 appPrefsWrapper,
@@ -106,101 +93,6 @@ class ReaderPostListViewModelTest {
                 readerTracker,
                 siteStore
         )
-        val observable = viewModel.getNewsDataSource()
-        observable.observeForever(observer)
-    }
-
-    @Test
-    fun `when view model starts pull is invoked`() {
-        viewModel.start(initialTag, false, false)
-        verify(newsManager).pull(false)
-    }
-
-    @Test
-    fun `view model propagates news items`() {
-        viewModel.start(initialTag, false, false)
-        liveData.postValue(item)
-        liveData.postValue(null)
-        liveData.postValue(item)
-
-        val inOrder = inOrder(observer)
-        inOrder.verify(observer).onChanged(item)
-        inOrder.verify(observer).onChanged(null)
-        inOrder.verify(observer).onChanged(item)
-    }
-
-    @Test
-    fun `view model propagates dismiss to NewsManager`() {
-        viewModel.onNewsCardDismissed(item)
-        verify(newsManager).dismiss(item)
-    }
-
-    @Test
-    fun `when view model starts emits null on first onTagChanged`() {
-        viewModel.start(otherTag, false, false)
-        // propagates the item since the card hasn't been shown yet
-        liveData.postValue(item)
-        viewModel.onTagChanged(initialTag)
-        // the card has been shown for the initialTag
-        viewModel.onNewsCardShown(item, initialTag)
-        // propagates null since the card should be visible only for initialTag
-        viewModel.onTagChanged(otherTag)
-        val inOrder = inOrder(observer)
-        inOrder.verify(observer).onChanged(item)
-        inOrder.verify(observer).onChanged(null)
-    }
-
-    @Test
-    fun `news item is available only for first tag for which card was shown`() {
-        viewModel.start(otherTag, false, false)
-        // propagate the item since the card hasn't been shown yet
-        liveData.postValue(item)
-        viewModel.onTagChanged(initialTag)
-        // the card has been shown for the initialTag
-        viewModel.onNewsCardShown(item, initialTag)
-        viewModel.onTagChanged(otherTag)
-        // do not propagate the item since the card should be visible only for initialTag
-        liveData.postValue(item)
-        // do not propagate the item since the card should be visible only for initialTag
-        liveData.postValue(item)
-        // do not propagate the item since the card should be visible only for initialTag
-        liveData.postValue(item)
-        // do not propagate the item since the card should be visible only for initialTag
-        liveData.postValue(item)
-        // propagate the item since the initialTag is selected
-        viewModel.onTagChanged(initialTag)
-        val inOrder = inOrder(observer)
-        inOrder.verify(observer).onChanged(item)
-        inOrder.verify(observer).onChanged(null)
-        inOrder.verify(observer).onChanged(item)
-    }
-
-    @Test
-    fun `view model propagates dismiss to NewsTracker`() {
-        viewModel.onNewsCardDismissed(item)
-        verify(newsTracker).trackNewsCardDismissed(argThat { this == READER }, any())
-    }
-
-    @Test
-    fun `view model propagates CardShown to NewsTracker`() {
-        whenever(newsTrackerHelper.shouldTrackNewsCardShown(any())).thenReturn(true)
-        viewModel.onNewsCardShown(item, initialTag)
-        verify(newsTracker).trackNewsCardShown(argThat { this == READER }, any())
-        verify(newsTrackerHelper).itemTracked(any())
-    }
-
-    @Test
-    fun `view model does not propagates CardShown to NewsTracker`() {
-        whenever(newsTrackerHelper.shouldTrackNewsCardShown(any())).thenReturn(false)
-        viewModel.onNewsCardShown(item, initialTag)
-        verify(newsTracker, times(0)).trackNewsCardShown(argThat { this == READER }, any())
-        verify(newsTrackerHelper, times(0)).itemTracked(any())
-    }
-
-    @Test
-    fun `view model propagates ExtendedInfoRequested to NewsTracker`() {
-        viewModel.onNewsCardExtendedInfoRequested(item)
-        verify(newsTracker).trackNewsCardExtendedInfoRequested(argThat { this == READER }, any())
     }
 
     @Test
